fix: build traveller page
