+++ apps/web/src/app/[lang]/app/[type]/operations/details/page.tsx
@@ -28,7 +28,7 @@ import { $UniRefund_TravellerService_Travellers_TravellerListProfileDto } from "
 import { getBaseLink } from "src/utils";
 import { dataConfigOfParties } from "../../parties/table-data";
 import { getMerchantsApi } from "../../../actions/CrmService/actions";
-import { travellerTableSchema } from "../../parties/traveller/page";
+import { travellerTableSchema } from "../../parties/traveller/utils";
 import { getTravellers } from "../../../actions/TravellerService/actions";
 import { getMerchants, getSummary, getTags } from "./actions";
 import type { TaxFreeTag } from "./data";
