+++ apps/web/src/app/[lang]/app/[type]/parties/traveller/page.tsx
@@ -1,20 +1,10 @@
 "use server";
 import type { GetApiTravellerServiceTravellersData } from "@ayasofyazilim/saas/TravellerService";
-import { $UniRefund_TravellerService_Travellers_TravellerListProfileDto } from "@ayasofyazilim/saas/TravellerService";
 import type { ColumnFilter } from "@repo/ayasofyazilim-ui/molecules/tables";
 import TableComponent from "@repo/ui/TableComponent";
 import { getResourceData } from "src/language-data/TravellerService";
 import { tableFetchRequest } from "../../../actions/table-utils";
-
-export const travellerTableSchema = {
-  excludeList: [
-    "id",
-    "userAccountId",
-    "residenceCountryCode2",
-    "nationalityCountryCode2",
-  ],
-  schema: $UniRefund_TravellerService_Travellers_TravellerListProfileDto,
-};
+import { travellerTableSchema } from "./utils";
 
 export default async function Page({ params }: { params: { lang: string } }) {
   const { languageData } = await getResourceData(params.lang);
