+++ apps/web/src/app/[lang]/app/[type]/parties/traveller/utils.ts
@@ -1,0 +1,11 @@
+import { $UniRefund_TravellerService_Travellers_TravellerListProfileDto } from "@ayasofyazilim/saas/TravellerService";
+
+export const travellerTableSchema = {
+  excludeList: [
+    "id",
+    "userAccountId",
+    "residenceCountryCode2",
+    "nationalityCountryCode2",
+  ],
+  schema: $UniRefund_TravellerService_Travellers_TravellerListProfileDto,
+};
